requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.2.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
googlesearch-python==2020.0.2
html5lib==1.1
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
requests==2.24.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
terminado==0.9.1
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.25.9
wcwidth==0.2.5
webdriver-manager==3.2.2
webencodings==0.5.1
